requirements.txt
arrow==0.16.0
autopep8==1.5.4
click==7.1.2
flake8==3.8.3
Flask==1.1.2
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
prometheus-client==0.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
PyYAML==5.3.1
six==1.15.0
tencentcloud-sdk-python-intl-en==3.0.81
toml==0.10.1
Werkzeug==1.0.1
zipp==3.1.0
